+++ core/launcher/args/kurtosis_backend_config/kubernetes_backend_config.go
@@ -7,5 +7,9 @@ package kurtosis_backend_config
 
 type KubernetesBackendConfig struct {
 	StorageClass string `json:"storageClass"`
+
+
+	// TODO Remove this and replace wtih FilesArtifactExpansionVolumeSizeInMegabytes, because the API container
+	//  actually doesn't need enclave data volume size (because it doesn't create enclaves)
 	EnclaveSizeInMegabytes uint `json:"enclaveSizeInMegabytes"`
 }